requirements.txt
numpy
mss
opencv-python
pillow
psutil
pywin32
json
